Suppression route util/user et factorisation du code (#605)
