+++ frontend/app/utils/utils.ts
@@ -157,151 +157,3 @@ export class Utils {
     return strNew;
   }
 }
-
-export function buildObjectResolvePropertyProcessing(
-  data,
-  fieldsConfig,
-  moduleCode,
-  _objService,
-  _cacheService
-): Observable<any> {
-  /**
-   * Traite et résout les propriétés d'un ensemble de données en fonction des types de champs définis dans la configuration.
-   *   La résolution consiste à transformer la valeur retournée par l'api par celle d'affichage
-   *
-   *
-   * @param data - Données à traiter.
-   * @param fieldsConfig - Configuration des champs permettant la résolution de chaque propriété.
-   * @param moduleCode - Le code de module courrant
-   * @param _objService - Service utilisé pour la résolution des propriétés d'objet.
-   * @param _cacheService - Service utilisé pour la mise en cache des propriétés résolues.
-   * @returns Un observable émettant l'objet de données avec les propriétés résolues.
-   */
-
-  const dataProcessing$ =
-    data &&
-    data.items &&
-    data.items.length > 0 &&
-    fieldsConfig &&
-    Object.keys(fieldsConfig).length > 0
-      ? forkJoin(
-          data.items.map((dataItem) => {
-            const propertyObservables = {};
-            for (const attribut_name of Object.keys(fieldsConfig)) {
-              // si des données sont contenues dans dataItem.data merge avec dataItem
-              // cas des propriétés supplémentaires des visites
-              // TODO reflechir si on garde cette propriété ou si on met tout à plat dans dataItem
-              if (dataItem.data) {
-                dataItem = { ...dataItem, ...dataItem.data };
-              }
-              if (dataItem.hasOwnProperty(attribut_name)) {
-                propertyObservables[attribut_name] = resolveProperty(
-                  _objService,
-                  _cacheService,
-                  moduleCode,
-                  fieldsConfig[attribut_name],
-                  dataItem[attribut_name]
-                );
-              }
-            }
-            if (Object.keys(propertyObservables).length === 0) {
-              return of(dataItem);
-            }
-            return forkJoin(propertyObservables).pipe(
-              map((resolvedProperties) => {
-                const updatedSiteGroupItem = { ...dataItem };
-                for (const attribut_name of Object.keys(resolvedProperties)) {
-                  updatedSiteGroupItem[attribut_name] = resolvedProperties[attribut_name];
-                }
-                return updatedSiteGroupItem;
-              })
-            );
-          })
-        ).pipe(
-          map((resolvedSiteGroupItems) => ({
-            ...data,
-            items: resolvedSiteGroupItems,
-          }))
-        )
-      : of(data);
-  return dataProcessing$;
-}
-
-export function resolveProperty(
-  _objService,
-  _cacheService,
-  moduleCode,
-  elem,
-  val
-): Observable<any> {
-  if (elem.type_widget === 'date' || (elem.type_util === 'date' && val)) {
-    val = Utils.formatDate(val);
-  }
-  if (elem.type_util === 'types_site') {
-    val = val.map((item) => {
-      return item.label;
-    });
-  }
-  const fieldName = _objService.configUtils(elem, moduleCode);
-  if (val && fieldName && elem.type_widget) {
-    return getUtil(_cacheService, elem.type_util, val, fieldName, elem.value_field_name);
-  }
-
-  return of(val);
-}
-
-function getUtil(
-  _cacheService,
-  typeUtil: string,
-  id,
-  fieldName: string,
-  idFieldName: string | null = null
-) {
-  if (Array.isArray(id)) {
-    return getUtils(_cacheService, typeUtil, id, fieldName, idFieldName);
-  }
-
-  var urlRelative = `util/${typeUtil}/${id}`;
-
-  if (idFieldName) {
-    urlRelative += `?id_field_name=${idFieldName}`;
-  }
-
-  const sCachePaths = `util|${typeUtil}|${id}`;
-
-  return _cacheService.cache_or_request('get', urlRelative, sCachePaths).pipe(
-    mergeMap((value) => {
-      let out;
-      if (fieldName === 'all') {
-        out = value;
-      } else if (fieldName.split(',').length >= 2) {
-        for (const fieldNameInter of fieldName.split(',')) {
-          if (value[fieldNameInter]) {
-            out = value[fieldNameInter];
-            break;
-          }
-        }
-      } else {
-        out = value[fieldName];
-      }
-      return of(out);
-    })
-  );
-}
-
-function getUtils(_cacheService, typeUtilObject, ids, fieldName, idFieldName) {
-  if (!ids.length) {
-    return of(null);
-  }
-  const observables: any[] = [];
-
-  for (const id of ids) {
-    observables.push(getUtil(_cacheService, typeUtilObject, id, fieldName, idFieldName));
-  }
-
-  return forkJoin(observables).pipe(
-    concatMap((res) => {
-      return of(res.join(', '));
-    })
-  );
-}
